refactor(kitchenpos): move JpaRepository from infra to domain
