+++ src/main/java/kitchenpos/menugroup/domain/JpaMenuGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.menugroup.infra.repository;
+package kitchenpos.menugroup.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
